+++ server/utils/github/organization.py
@@ -21,23 +21,6 @@ class GitHubAppOrg(BaseGitHubApp):
             auth_type="install_token",
         )
 
-    def get_repo_collaborators(self, repo_name: str, owner_name: str) -> list | None:
-        """Get repo collaborators.
-
-        Args:
-            repo_name (str): The name of the repo.
-            owner_name (str): The name of the owner.
-
-        Returns:
-            list: The repo collaborators.
-        https://docs.github.com/zh/rest/collaborators/collaborators?apiVersion=2022-11-28#list-repository-collaborators
-        """
-
-        return self.base_github_rest_api(
-            f"https://api.github.com/repos/{owner_name}/{repo_name}/collaborators",
-            auth_type="install_token",
-        )
-
     def get_org_members(self, org_name: str) -> list | None:
         """Get a list of members of an organization.
